+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/constants
 
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- beta tag
+
 ## 1.0.0-beta.4
 
 ### Patch Changes