+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/shared
 
+## 1.0.0-beta.4
+
+### Patch Changes
+
+- beta tag
+- Updated dependencies
+  - @pipelab/migration@1.0.0-beta.4
+
 ## 1.0.0-beta.3
 
 ### Patch Changes