+++ plugins/plugin-discord/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pipelab/plugin-discord
 
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- beta tag
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.0-beta.6
+  - @pipelab/shared@1.0.0-beta.4
+
 ## 1.0.0-beta.5
 
 ### Patch Changes